+++ static/attraction.html
@@ -162,5 +162,6 @@
     <div class="copyright">COPYRIGHT © 2021 台北一日遊</div>
     <!-- copyright結束 -->
     <script src="/static/js/script_attraction.js"></script>
+    <script src="/static/js/login-functions.js"></script>
   </body>
 </html>
