+++ static/index.html
@@ -121,6 +121,7 @@
     <!-- copyright結束 -->
 
     <script src="/static/js/script.js"></script>
+    <script src="/static/js/login-functions.js"></script>
     <!-- 引入外部JavaScript文件 -->
   </body>
 </html>
